refactor: rename package to upstream identifier
