+++ plugin/upstream_identifier/evaluator/evaluator_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/goto/salt/log"
 	"github.com/stretchr/testify/assert"
 
-	"github.com/goto/optimus/plugin/upstream_generator/evaluator"
+	"github.com/goto/optimus/plugin/upstream_identifier/evaluator"
 )
 
 func TestNewEvaluatorFactory(t *testing.T) {
